jsdialog: fix updating on container "reorder_child" or "move"
The accessibility checker sidebar did not update properly when
new issues were added or were removed. The accessiblity checker
is adding and removing new widget parts (one for each new issues)
to the grids, which is using "reorder_child" and "move" methods.
If we call "reorder_child" with update, it fails in online as it
doesn't find the parent "box" ("box_document" for example) because
the id is not set properly. Doing a full update on "reorder_child"
fixes the updating problem.

Change-Id: Ia3032516a7cee0271f8b05b22e13dde2b7430e66
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/166962
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins