makefile simplification: replace $(call gb_CustomTarget_get_workdir,foo)
…by a simple/static $(gb_CustomTarget_workdir)/foo

The build system has a lot of overly complicated leftovers from when it
was introduced and had not only deal with split repositories but also
had to coexist with another buildsystem. Along with lots of copy'n'paste
along the years the makefiles became hard to grasp for newcomers with
all our calls and evals.
As a first step to streamline that, the macros from TargetLocations that
simply prefix a static path to the argument (and similar of the same
kind) are a natural pick before simplifying the rules themselves/getting
rid of a bunch of eval statements.

(cherry-picked from 0c4c84a14b01c71c76a9c45a7f26aec4d64f3e4f)
Conflicts:
	cli_ure/CustomTarget_cli_ure_assemblies.mk
	extras/CustomTarget_tplwizard.mk
	odk/CustomTarget_doxygen.mk
	sw/CustomTarget_source.mk
	vcl/CustomTarget_qt5_moc.mk
	vcl/CustomTarget_qt6_moc.mk

Change-Id: Ia06dbbcd5d1994755a2ff05b84f72ccbc4e3cab5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167005
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>