Revert "tdf#150037 Writer: text fieldmark behaviour changed"
This reverts 24.8 commit 96323a10d3a55d212c350886e2a1344c0cd2ba95.
    A mouse click selects the fieldmark to be able to overwrite
    the field with the next text input.
But I did not revert:
    Additionally the EN SPACE (0x2002) character is visualized
    by the DEGREE symbol when 'View/Formatting Marks' is active.

The problem was that the field itself was also selected,
so not only was the placeholder text overwritten,
but so was the entire field.

Additionally, it is ONLY placeholder text that should be selected,
but this was also selecting non-placeholder
(i.e. custom user-entered text) that was being pre-selected
which is NOT what happens in MS Word.

The whole point of the patch in the first place was to
emulate MS Word's response to placeholder text,
so we should also be emulating what happens
to non-placeholder text. That would require tracking
whether the field's text was placeholder or not,
which is not currently track in legacy fields.

Change-Id: I0a1a22f82827cd5f9d9abfb52cd562dff4651b29
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167139
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins