The previous behavior of using the TerminationVetoException has changed
"Since LibreOffice 5.3: Throwing this exception will only prevent termination. Exiting LibreOffice will close all the windows but the process will keep running."

Currently, the behavior is different, because not only is LibreOffice
prevented from terminating, but the last document cannot be closed either. The expected behavior is all windows should be closed, but the
process should be kept running when using a TerminationVetoException.

Change-Id: I8764aefccc4c9feec3ee0e45a54bf0aebcf16c5d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167194
Reviewed-by: Samuel Mehrbrodt <samuel.mehrbrodt@allotropia.de>
Tested-by: Jenkins