tdf#160806 tdf#160837 gtk3 a11y: Port from deprecated atk_focus_tracker_notify
`atk_focus_tracker_notify` takes care of sending an
"AtkObject:focus-event" event for the given object.
However, both, `atk_focus_tracker_notify` and the "AtkObject:focus-event"
ATK event (which triggers a "focus" AT-SPI event) are
deprecated, quoting [1]:

> Focus tracking has been dropped as a feature to be implemented by ATK
> itself. As AtkObject::focus-event was deprecated in favor of a
> AtkObject::state-change signal, in order to notify a focus change on
> your implementation, you can use atk_object_notify_state_change()
> instead.

Move the existing call to `atk_object_notify_state_change` up and
use instead of the deprecated function.

For the example of the "Print" button in
Writer's standard toolbar, it can now be
seen that a "object:state-changed:focused"
instead of a "focus" event is received with
a corresponding listener when tabbing to
the button and then further:

Before:

    88.7 focus:(0, 0, 0)
            source: [push button | Print]
            application: [application | soffice]
    88.7 object:state-changed:focused(1, 0, 0)
            source: [push button | Print]
            application: [application | soffice]
    89.4 object:state-changed:focused(0, 0, 0)
            source: [push button | Print]
            application: [application | soffice]

After:

    80.0 object:state-changed:focused(1, 0, 0)
            source: [push button | Print]
            application: [application | soffice]
    80.0 object:state-changed:focused(1, 0, 0)
            source: [push button | Print]
            application: [application | soffice]
    80.8 object:state-changed:focused(0, 0, 0)
            source: [push button | Print]
            application: [application | soffice]

The first event is the one emitted fromt this place,
the others are from elsewhere.

This also already shows that duplicate events are
sent. Eliminating those which also cause
other problems is planned in upcoming steps.

[1] https://docs.gtk.org/atk/func.focus_tracker_notify.html

Change-Id: Ieb9d25445b620681b01b887c0c120c80a43379ef
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167207
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins