gtk3 a11y: Drop VclEventId::ToolboxHighlight{,Off} processing
Drop the manual processing of `VclEventId::ToolboxHighlight`
in the gtk3 a11y bridge.

As mentioned in

    commit c45c64aeb57dce91965d7be54601128946455c90
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon May 6 14:44:36 2024 +0200

        tdf#160806 tdf#160837 gtk3 a11y: Drop handling of some VclEventIds

, the a11y bridge should not have to handle VCL events itself,
but all relevant events should be forwarded to the a11y
event listener as appropriate `AccessibleEventObject` and be handled
by the a11y event listener instead, see `AtkListener::notifyEvent`
for the gtk3 implementation.

In a quick test, toolbar items were still announced
just fine by the Orca screen reader when using the
gtk3 VCL plugin with this commit in place.

`VCLXAccessibleToolBox::ProcessWindowEvent` handles
`VclEventId::ToolboxHighlight` and
`VclEventId::ToolboxHighlightOff` whose manual processing
is now dropped from gtk3, so the former might be a good place
to further look into if anything is still missing after
this commit.

The now dropped handling was originally added in

    commit e9816343cd71dfe556e6ace12d67c642a8dde635
    Author: Jens-Heiner Rechtien <hr@openoffice.org>
    Date:   Thu May 11 12:32:22 2006 +0000

        INTEGRATION: CWS atkbridge2 (1.2.4); FILE MERGED
        2006/05/08 13:45:15 obr 1.2.4.1: #135353# fixed crashes when closing sub-toolbars and application exit

    Change-Id: Ib6cfafbc3711fce6a48120439654dc6e6c15971a

Change-Id: I87061fdc595a0a29c2a3dce88ab13c9d7d7ba1b0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167253
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>