gtk3 a11y: Drop VclEventId::MenuHighlight processing
Drop the manual processing of `VclEventId::MenuHighlight`
in the gtk3 a11y bridge.

As mentioned in

    commit c45c64aeb57dce91965d7be54601128946455c90
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon May 6 14:44:36 2024 +0200

        tdf#160806 tdf#160837 gtk3 a11y: Drop handling of some VclEventIds

, the a11y bridge should not have to handle VCL events itself,
but all relevant events should be forwarded to the a11y
event listener as appropriate `AccessibleEventObject` and be handled
by the a11y event listener instead, see `AtkListener::notifyEvent`
for the gtk3 implementation.

In a quick test, menu items were still announced
just fine by the Orca screen reader when using the
gtk3 VCL plugin with this commit in place.

By now, most menus should be native GTK widgets anyway.
For non-native menus,
`OAccessibleMenuBaseComponent::ProcessMenuEvent` contains
handling for the `VclEventId::MenuHighlight` event that
should be taking care of forwarding that as the corresponding
a11y event to the a11y bridges. If anything is missing after
this commit, that's probably a good starting point for further
analysis.

Change-Id: Ia8686101dc559ed98b48285eac15c35fc10818af
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167254
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins