gtk3 a11y: Drop VclEventId::TabpageActivate processing
Drop the manual processing of `VclEventId::TabpageActivate`
in the gtk3 a11y bridge.

As mentioned in

    commit c45c64aeb57dce91965d7be54601128946455c90
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon May 6 14:44:36 2024 +0200

        tdf#160806 tdf#160837 gtk3 a11y: Drop handling of some VclEventIds

, the a11y bridge should not have to handle VCL events itself,
but all relevant events should be forwarded to the a11y
event listener as appropriate `AccessibleEventObject` and be handled
by the a11y event listener instead, see `AtkListener::notifyEvent`
for the gtk3 implementation.

By now, most if not all tab pages should be native GTK widgets
anyway.
For non-native ones, `VCLXAccessibleTabControl::ProcessWindowEvent`
contains handling for the `VclEventId::TabpageActivate` event that
should be taking care of forwarding that as the corresponding
a11y event to the a11y bridges. If anything is still missing after
this commit, that's probably a good starting point for further
analysis.

Change-Id: I5b3f64c973d4030d2b68be6dff1d771bc0cef405
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167255
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>