gtk3 a11y: Drop WindowEventHandler completely
As mentioned in

    commit c45c64aeb57dce91965d7be54601128946455c90
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon May 6 14:44:36 2024 +0200

        tdf#160806 tdf#160837 gtk3 a11y: Drop handling of some VclEventIds

, the a11y bridge should not have to handle VCL events itself,
but all relevant events should be forwarded to the a11y
event listener as appropriate `AccessibleEventObject` and be handled
by the a11y event listener instead, see `AtkListener::notifyEvent`
for the gtk3 implementation.

Now that handling of all VCL events has been dropped in
previous commits, drop the `WindowEventHandler` completely.
This also means that the call to `InitAtkBridge` when initializing
the `GtkInstance` and the `DeInitAtkBridge` on destruction
is no longer necessary.

In a quick test with the Orca screen reader and the gtk3
VCL plugin, nothing suspicious showed up, things were
still working as expected and experienced prior to the
change series dropping the VCL event handling from the
gtk3 a11y bridge.

Change-Id: Ieed270b636f361c57cd4e78aceaecc7d174409d6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167256
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins