tdf#160971 gtk3 a11y: Set role for custom editable combobox
Set the combobox a11y role for the box in the .ui file
used for the custom gtk3 combobox implementation.

With this in place, moving focus to the "Font Name" or
another editable combobox in the Writer toolbar now
makes Orca announce that one as "editable combobox"
rather than just "text".

See also

    commit 1e851093f0148d2c55fc3fd377d274f6703c71c9
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Apr 18 14:02:25 2024 +0200

        tdf#159910 gtk3 a11y: Keep a11y props for combobox

for a very similar change that caused the proper
role to be announced for non-editable comboboxes.

Change-Id: I3062e5c062911040ef0ec5721d15fee78d85d0dd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167262
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>