gtk3 a11y: No longer manually add focused state for focus object
No longer manually add `ATK_STATE_FOCUSED` to the
ATK focus object.

The focused object should should have the
`css::accessbibility::AccessibleStateType::FOCUSED`
state set on UNO level, in which case this is already
mapped to the focused state in the lines above the
newly removed code, so explicitly addding it
again shouldn't be necessary.

If the state is missing, that most likely needs
to be fixed in the underlying `XAccessibleStateSet`
implementation.

Therefore, remove what looks like a workaround.
If any issues show up after this, let's rather take
a look at the underlying issues and try to fix things
there.

Change-Id: Ie7252c4b2dd2ec15c7a3bbc6a1a5e2830cdd39a7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167309
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins