svx: a11y check entry box - provide accessible name and description
If the accessible name and description is not set in the .ui file,
the code will look into help for accessible description, which will
always return an empty string as the widgets are added dynamically.
Searching in help takes some time (2ms), which can add up and is
"felt" in accessibility check sidebar in online when the documents
has a lot of accessibility issues (it can add up to seconds). To
avoid this cost and for better accessibility info, it is better
to provide some general description for this widgets.

Change-Id: I91c5a091781532068c517be937d225fc9345dc5c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167346
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>
(cherry picked from commit 31b3d5a1062366a7ab0cf0fa42359b90d1ed3c4b)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167316
Tested-by: Jenkins