lok: don't try to search for legacy label from/to
Searching for legacy label from/to can take quite some time in
complex hierarchies (wallclock timed to 100ms) and this quickly
can add up. In case of accessibility checker sidebar with many
entries it could add up to 10s or more (depending on the amount of
accessibility issues found in the document).

The legacy label from/to is not used when we load from .ui file so
it is not relevant for online currently. It is only relevant for ui
found in extensions or starbasic ui.

Also avoid making a copy of the vector of mnemonic label for no
particular reason.

Change-Id: I72002e6440a06f6e8548b7bddb46c3ab283ea283
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167347
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
(cherry picked from commit 1353ae001ed83c1f20b81df1708f694427c587a4)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167315
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>