Revert "tdf#160706 speed up loading conditional formatting rule in XLS (3)"
This reverts commit e1268721d4266ed1254da99581a7c2c3dfb72ba6.

caolanm:

With the old "bForceName" of true then previously if the style name existed it was renamed to something else and then a fresh style with the requested name created. While now we just reuse it.

I'm not very clued in to what calc does, but in writer we might do something like this if we are inserting into an existing document and don't want to clobber existing styles. Except here we sort of did clobber existing styles by renaming them. So I'm unsure of what on earth the original idea was.

So I wonder about a "insert into existing calc document" scenario?

noelg:

Caolan, I was operating under the assumption that this comment here is the desired behaviour.
But now that I read it again, I am not so sure.
This commit should probably be backed out.

Change-Id: If503efa9fa1a7886eb57bb62d1327f81e4410981
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167318
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins