tdf#160971 a11y: Send full text on changed combobox text
When the text of an editable combobox is changed
(e.g. using the up/down keys), the Orca screen
reader on Linux announces the newly inserted text.

This has been the case for other GTK or Qt
applications for a while, and with recent Orca
commit [1]

    commit 3a9e6b8d7b16bf2fc7919868cfd1a16e44422710
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri May 10 10:16:58 2024 +0200

        soffice: Use default logic for editable combobox value change

, the same logic is used for LibreOffice as well.

For the gtk3 VCL plugin which has a custom combobox
implementation using native GTK widgets, this generally
works since

    commit 9f078ed7b625e86182d64d5ccfbb410cdd38081c
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue May 7 10:04:16 2024 +0200

        tdf#160971 gtk3 a11y: Set role for custom editable combobox

However, the qt6 VCL plugin uses the VCL combobox
implementation, and only the actual difference between
the text of the previous and current entry was sent
in the TEXT_CHANGED event, resulting in Orca only
announcing those letters that were added/changed,
e.g. just "ans Narrow" when changing the font in the
Writer formatting toolbar from "Liberation Serif"
to "Liberation Sans Narrow".

This doesn't really make clear what entry is selected.
Align the a11y event with what GTK and Qt do
and set the full old and new entry texts in the event.

To do that, add a new virtual
`VCLXAccessibleTextComponent:PreferFullTextInTextChangedEvent`
that defaults to false to keep the previous behavior as default,
and override it for `VCLXAccessibleEdit` to return true
in the case that the edit is the subedit of a combobox
(the parent has a combobox role).
Use this in `VCLXAccessibleTextComponent::SetText` to
determine whether to notify just of the changed characters
or to send the whole old/new text.

With this in place, Orca also announces the whole new
entry text (e.g. "Liberation Sans Narrow" for the above
example) when using the qt6 VCL plugin.
(It currently additionally announces an extra "Selection
deleted" when switching entries, as the text selection
also changes, but that aspect is to be handled separate
from this change here.)

[1] https://gitlab.gnome.org/GNOME/orca/-/commit/3a9e6b8d7b16bf2fc7919868cfd1a16e44422710

Change-Id: I240aa0ad5ac9585e007d67a8c69e305cf1f38185
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167479
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>