tdf#160824 vcl: Require mouse over spinfield to mouse-wheel through values
In the same way that

    commit 22250df05830700b2555348b8ac46ee1007d0e5d
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Dec 8 11:50:40 2023 +0100

        tdf#158548 vcl: Require mouse over listbox to mouse-wheel through entries

restricted changing listbox values for the focused
listbox on mouse-wheel to the case when the listbox
is also hovered over by the mouse, do the same for
`SpinField` as well, to avoid accidently changing
the value as described e.g. for the tdf#160824
scenario.

This is also in line with Qt spin boxes that
only change their values on mouse-wheel when
the mouse pointer is above them.

For the gtk3 VCL plugin that uses native spin boxes,
this commit has no effect, and the value doesn't
change on mouse-wheel at all.

Change-Id: I22277294f76655cd13b1dc57416c82afe44460d1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167604
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins