sw: make testTdf131963 less strict
Asserting number of pages is rather questionable because any number of
things can change that and you will spend days debugging that.

For some reason this test has 12 pages on my libreoffice-24-2 branch
build and 11 pages on my master build, either of which are clearly
better than the 2 pages result of the bug that was fixed there.

Change-Id: Ie1e97d5cc973f317a2d6ca8a7d13822385b134f4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167627
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>