tdf#69192 fix cmdlinehelp: --show[=slide_number], etc.
I had skimmed the help file earlier, and concluded that { }
are used elsewhere to indicate optional paramaters.
However, there is --quickstart[=no], and -env:<VAR>[=<VALUE>]
which exactly match my optional addition.

Thanks Stéphane for pointing this out.
Also included are some other changes we made
to clarify the help wording in general,
so the documentation and the cmdline help are in sync.

We dropped
    Files other than Impress documents are opened in
    default mode, regardless of previous mode.
because:
-from an initial reading it doesn't convey any useful meaning.
-searching for other uses of "mode" in the file suggests it means
 that prior --print/--convert-to/--view modes
 no longer apply. This override of prior modes is true for all modes,
 but only this argument tries to spell this out.
-it seems rather inconceivable that any user would attempt
 to mix --show with any other arguments or more than one file.

Change-Id: I191e572506cf2b39ed6c9ff4ebaf66f258a176a1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/167633
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>